Edit component names and organize folder structure
