+++ src/app/document/[docId]/page.tsx
@@ -1,6 +1,6 @@
 import { notFound } from 'next/navigation';
 import { use } from 'react';
-import PdfViewer from '@/components/pdf-viewer';
+import PdfViewer from '@/components/features/document/pdf-viewer';
 import { bucketName } from '@/lib/supabase';
 
 type DocParams = Promise<{ docId: string }>;
