+++ src/app/mypage/cart/page.tsx
@@ -1,6 +1,6 @@
 'use client';
-import CartList from '@/components/cart-list';
-import PaymentSummary from '@/components/payment-summary';
+import CartList from '@/components/features/mypage/cart/cart-list';
+import PaymentSummary from '@/components/features/mypage/cart/payment-summary';
 import { CartItem } from '@/types/my-card';
 import { useState } from 'react';
