+++ src/app/mypage/layout.tsx
@@ -1,6 +1,6 @@
 import '@/app/globals.css';
 import { UserProvider } from '../context/user-context';
-import MyPage from '@/components/my-page';
+import MyPage from '@/components/features/mypage/my-page';
 
 export default function RootLayout({
   children
