+++ src/components/common/list-header.tsx
@@ -1,7 +1,7 @@
 'use client';
 import * as React from 'react';
 import { useState, useEffect } from 'react';
-import { Button } from './ui/button';
+import { Button } from '../ui/button';
 import {
   Carousel,
   CarouselApi,
