+++ src/components/features/course/course-header.tsx
